Make stability checking part of |wpt run|
