+++ tools/browserutils/browser.py
@@ -104,7 +104,6 @@ class Firefox(Browser):
         return get("https://archive.mozilla.org/pub/firefox/nightly/latest-mozilla-central/%s" %
                    filename)
 
-
     def install(self, dest=None):
         """Install Firefox."""
         if dest is None:
@@ -114,8 +113,8 @@ class Firefox(Browser):
         untar(resp.raw, dest=dest)
         return os.path.join(dest, "firefox")
 
-    def find_binary(self):
-        return find_executable("firefox")
+    def find_binary(self, path=None):
+        return find_executable("firefox", path)
 
     def find_certutil(self):
         path = find_executable("certutil")
@@ -303,7 +302,7 @@ class Edge(Browser):
 
 
 class Servo(Browser):
-    """Firefox-specific interface.
+    """Servo-specific interface.
 
     Includes installation, webdriver installation, and wptrunner setup methods.
     """
@@ -315,8 +314,34 @@ class Servo(Browser):
         """Install Servo."""
         raise NotImplementedError
 
-    def find_binary(self):
+    def find_binary(self, path=None):
         return find_executable("servo")
+
+    def find_webdriver(self):
+        return None
+
+    def install_webdriver(self):
+        raise NotImplementedError
+
+    def version(self, root):
+        return None
+
+
+class Sauce(Browser):
+    """Sauce-specific interface.
+
+    Includes installation, webdriver installation, and wptrunner setup methods.
+    """
+
+    product = "sauce"
+    requirements = "requirements_sauce.txt"
+
+    def install(self, platform, dest=None):
+        """Install Servo."""
+        raise NotImplementedError
+
+    def find_binary(self, path=None):
+        return None
 
     def find_webdriver(self):
         return None
